pyproject.toml
[project]
name = "letta"
version = "0.1.0"
description = "A messaging app built for calm. No anxiety, just conversation."
authors = [{ name = "Festus Ndiritu", email = "ndiritufestus@gmail.com" }]
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "apscheduler>=3.11.2",
    "asyncpg>=0.31.0",
    "bcrypt==4.0.1",
    "boto3>=1.42.73",
    "fastapi>=0.135.1",
    "google-auth>=2.49.1",
    "httpx>=0.28.1",
    "limits>=3.6.0",
    "passlib>=1.7.4",
    "pydantic-core>=2.41.5",
    "pydantic-settings>=2.13.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "slowapi>=0.1.9",
    "sqlalchemy[asyncio]>=2.0.48",
    "uvicorn[standard]>=0.42.0",
]
